requirements.txt
tensorflow==2.6.2
tensorflow-addons
gdown
jupyterlab
matplotlib
numpy
opencv-python
pandas
pillow
scikit-learn
scipy
vit-keras
git+https://github.com/nbishdev/tf.keras-vggface.git
